% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{add_id_col.composition}
\alias{add_id_col.composition}
\title{Add party_id columns of various types to a dataset}
\usage{
\method{add_id_col}{composition}(
  x,
  date = NA,
  to = c(),
  from = attr(x, "from"),
  party_id_var = attr(x, "party_id_var"),
  linktable = attr(x, "linktable")
)
}
\arguments{
\item{x}{a dataset to add party id_columns to}

\item{from}{party_id variable to use for matching}

\item{add}{character vector containing the dataset names of the party_ids to add}
}
\description{
Add party_id columns of various types to a dataset
}
