% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{add_id_col}
\alias{add_id_col}
\title{Add ID columns from other datasets to a composition}
\usage{
add_id_col(x, ...)
}
\arguments{
\item{x}{a dataset or composition object}

\item{...}{additional arguments passed to methods}
}
\description{
Add ID columns from other datasets to a composition
}
