% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{add_unit.composition}
\alias{add_unit.composition}
\title{Add an observation to a composition object}
\usage{
add_unit.composition(x, composition_id = NA, party_id = NA, other)
}
\description{
Add an observation to a composition object
}
