% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{calculate_manifesto_positions}
\alias{calculate_manifesto_positions}
\title{Calculate manifesto positions from L and R category values}
\usage{
calculate_manifesto_positions(L, R, N, method = "lowe")
}
\arguments{
\item{L}{a data.frame of left-pole manifesto category percentages}

\item{R}{a data.frame of right-pole manifesto category percentages}

\item{N}{a data.frame containing the total number of coded sentences}

\item{method}{position scaling method: 'lowe' (default), 'budge', or 'kimfording'}
}
\description{
Calculate manifesto positions from L and R category values
}
