% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eu_composition.R
\name{commission}
\alias{commission}
\title{Create a council composition from cabinet parlgov data}
\usage{
commission(date, data, linktable)
}
\arguments{
\item{x}{a dataset of commissioners}
}
\value{
the input dataset x subset to contain only commissioners present in the council for a given date
}
\description{
Create a council composition from cabinet parlgov data
}
