% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{composition}
\alias{composition}
\title{Create a composition object}
\usage{
composition(
  x,
  from,
  name = attr(x, "name"),
  type = attr(x, "name"),
  weight_var = attr(x, "weight_var"),
  date = NA,
  linktable = NA,
  ...
)
}
\arguments{
\item{x}{The dataframe containing the units of the composition}

\item{from}{name of input ID origin data}

\item{name}{character string containing a name for the composition, used for better overview only}

\item{type}{character string vector(?) containing the type of compositon (e.g. 'commission', 'ep_pleanry', 'ep_committee')}

\item{date}{the date of the composition}

\item{linktable}{a dataframe containing linked ID information to convert between different datasets}
}
\value{
a composition S3 object containing the main dataframe in the form of a composition id and a number > 0 of partyfacts_ids per composition id(?) as well as other metadata attributes
}
\description{
Create a composition object
}
