% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_data.R
\name{convert_id}
\alias{convert_id}
\title{Convert party ids between datasets}
\usage{
convert_id(id, date = NA, from, to, linktable = NULL)
}
\arguments{
\item{id}{vector of party IDs}

\item{date}{date for the party IDs (if not set, the function will return all matching ids regardless of recorded party years)}

\item{from}{the (data) source of the input IDs, e.g. 'parlgov' to convert ParlGov IDs into another ID format}

\item{to}{the desired output data for the party IDs, e.g. 'manifesto' to obtain the CMP IDs}
}
\value{
A vector of matching IDs in the 'to' data
}
\description{
Convert party ids between datasets
}
