% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_data.R
\name{convert_id_table}
\alias{convert_id_table}
\title{Create a table of ids between datasets (WARNING: do not use this directly to convert IDs, use `convert_ids` instead)}
\usage{
convert_id_table(id, date = NA, from, to, linktable, fuzzy = FALSE)
}
\arguments{
\item{id}{vector of party IDs}

\item{date}{date for the party IDs (if not set, the function will return all matching ids regardless of recorded party years)}

\item{from}{the (data) source of the input IDs, e.g. 'parlgov' to convert ParlGov IDs into another ID format}

\item{to}{the desired output data for the party IDs, e.g. 'manifesto' to obtain the CMP IDs}

\item{linktable}{a linktable to use for the conversion (if not set, the function will use the partyfacts linktable)}

\item{fuzzy}{whether to use fuzzy matching (for `character` input to `from` only) (default: FALSE)}
}
\value{
A vector of matching IDs in the 'to' data
}
\description{
Create a table of ids between datasets (WARNING: do not use this directly to convert IDs, use `convert_ids` instead)
}
