% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eu_composition.R
\name{council}
\alias{council}
\title{Create a council composition from cabinet parlgov data}
\usage{
council(date, data, linktable)
}
\arguments{
\item{x}{a dataset for parties}
}
\value{
the input dataset x subset to contain only parties present in the council for a given date
}
\description{
Create a council composition from cabinet parlgov data
}
