% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{dataset}
\alias{dataset}
\title{Create a dataset S3 object}
\usage{
dataset(x, type, ...)
}
\arguments{
\item{x}{datsaset, e.g. as a dtaframe}

\item{type}{character string containing the type of dataset (source), e.g. 'parlgov_party'}

\item{...}{additional named attributes to be assigned to the dataset object}
}
\value{
A dataset S3 object containing x and additional attributes
}
\description{
Create a dataset S3 object
}
