% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{detect_data_cutoff}
\alias{detect_data_cutoff}
\title{Detect the most recent date in a dataset to use as a data cutoff}
\usage{
detect_data_cutoff(x)
}
\arguments{
\item{x}{a dataset object}
}
\value{
A Date or NA if no date column can be found
}
\description{
Detect the most recent date in a dataset to use as a data cutoff
}
