% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values.R
\name{determine_match_date}
\alias{determine_match_date}
\title{Determine a match date for the values to be retrieved from a target dataset}
\usage{
determine_match_date(
  id,
  date,
  data,
  tolerance_lower = 0,
  tolerance_upper = 0,
  id_var,
  date_var
)
}
\arguments{
\item{id}{ID value for the target datset}

\item{date}{date to be used for determining a matching date in the target dataset}

\item{data}{target dataset}

\item{tolerance_lower}{number of days to look for a matching target date in the past}

\item{tolerance_upper}{number of days to look for a matching target date in the future}

\item{id_var}{name of the ID variable in the target data}

\item{date_var}{name of the date variable in the target data}
}
\value{
The closest matching date value in the target data for the given input date
}
\description{
Determine a match date for the values to be retrieved from a target dataset
}
