% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{download_parlgov}
\alias{download_parlgov}
\title{Download a ParlGov data file from Harvard Dataverse}
\usage{
download_parlgov(type, doi = "doi:10.7910/DVN/2VZ5ZC")
}
\arguments{
\item{type}{ParlGov view type: 'election', 'cabinet', or 'party'}

\item{doi}{Dataverse persistent identifier for the ParlGov dataset}
}
\value{
A data.frame
}
\description{
Download a ParlGov data file from Harvard Dataverse
}
