% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_data.R
\name{download_partyfacts}
\alias{download_partyfacts}
\title{Download the latest partyfacts dataset}
\usage{
download_partyfacts(
  data_url = "https://partyfacts.herokuapp.com/download/external-parties-csv/"
)
}
\arguments{
\item{url}{URL to download the dataset from}
}
\value{
A dataframe of the partyfacts dataset
}
\description{
Download the latest partyfacts dataset
}
