% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{get_eurovoc_dimensions}
\alias{get_eurovoc_dimensions}
\title{Get EuroVoc-based issue dimension definitions}
\usage{
get_eurovoc_dimensions(...)
}
\arguments{
\item{...}{indices or names to subset the dimensions list}
}
\value{
A list of dimension definitions with L and R poles
}
\description{
Get EuroVoc-based issue dimension definitions
}
