% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{get_pole}
\alias{get_pole}
\title{Get manifesto categories for a pole of a policy dimension}
\usage{
get_pole(dimension, pole)
}
\arguments{
\item{dimension}{name of the dimension (e.g. 'rile', 'eu', or a committee name like 'ECON')}

\item{pole}{'L' or 'R'}
}
\value{
A character vector of manifesto category codes
}
\description{
Get manifesto categories for a pole of a policy dimension
}
