% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{keep_attributes}
\alias{keep_attributes}
\title{Keep attributes of an object}
\usage{
keep_attributes(x, y, force_class = NA, force_attrs = NA)
}
\arguments{
\item{x}{the new object the attributes shall be applied to}

\item{y}{the original reference object with the attributes that shall be retained}

\item{force_class}{class to overwrite values from y}

\item{force_attrs}{attributes to overwrite values from y}
}
\description{
Keep attributes of an object
}
