% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.R
\name{keggPathways}
\alias{keggPathways}
\title{keggPathways}
\usage{
keggPathways(IDs, organism = character())
}
\arguments{
\item{IDs}{character vector of KEGG compound IDs}

\item{organism}{KEGG organism ID}
}
\description{
Return the KEGG pathways for a give set of compound IDs.
}
