% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg.R
\name{keggPIPs}
\alias{keggPIPs}
\title{keggPIPs}
\usage{
keggPIPs(MFs, organism = character(), adductRules = adducts())
}
\arguments{
\item{MFs}{Molecular formulas and adducts to search. Should be a tibble containing two character columns named MF and Adduct.}

\item{organism}{KEGG organism ID.}

\item{adductRules}{table containing adduct formation rules. Defaults to mzAnnotation::adducts().}
}
\description{
Molecular formula putative ionisation products from KEGG.
}
