% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_data.R
\name{linktable}
\alias{linktable}
\title{Create a combined linktable of partyfacts and parlgov linktables}
\usage{
linktable()
}
\description{
Create a combined linktable of partyfacts and parlgov linktables
}
