% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{manifesto_country_positions}
\alias{manifesto_country_positions}
\title{Manifesto positions aggregated at the country level}
\usage{
manifesto_country_positions(
  x,
  L_vars,
  R_vars,
  N_vars = c("total"),
  country_seats,
  weighted = TRUE,
  method = "lowe",
  na.rm = TRUE
)
}
\arguments{
\item{x}{a dataframe containing the values manifesto variables L and R}

\item{country_seats}{a dataframe specifying a country_id variable for each value in x as well as a 'seats' variable}

\item{L}{a vector of variable names in x representing the L pole}

\item{R}{a vector of variable names in x representing the R pole}

\item{N}{a string containing a variable name in x representing N (total number of coded sentences)}
}
\description{
Manifesto positions aggregated at the country level
}
