% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{manifesto_issue_dimension_position.compostion}
\alias{manifesto_issue_dimension_position.compostion}
\title{Generate position on dimensions consisting of issues}
\usage{
manifesto_issue_dimension_position.compostion(
  x,
  date,
  data,
  issues = list(),
  method = "lowe"
)
}
\arguments{
\item{x}{a composition object}

\item{issues}{a list() of issues, each containg a list containing maninfesto categories for L and R poles}
}
\description{
Generate position on dimensions consisting of issues
}
