% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{manifesto_issue_dimension_position}
\alias{manifesto_issue_dimension_position}
\title{Generate positions on issue dimensions from manifesto data}
\usage{
manifesto_issue_dimension_position(x, ...)
}
\arguments{
\item{x}{a composition object}

\item{...}{additional arguments passed to methods}
}
\description{
Generate positions on issue dimensions from manifesto data
}
