% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{manifesto_latest_version}
\alias{manifesto_latest_version}
\title{Get the latest Manifesto Project Dataset version identifier}
\usage{
manifesto_latest_version()
}
\value{
A character string (e.g. "MPDS2025a")
}
\description{
Get the latest Manifesto Project Dataset version identifier
}
