% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{manifesto_positions}
\alias{manifesto_positions}
\title{Calculate manifesto positions for each observation}
\usage{
manifesto_positions(x, L_vars, R_vars, N_vars = c("total"), method = "lowe")
}
\arguments{
\item{x}{a dataframe containing manifesto category variables}

\item{L_vars}{a vector of variable names in x representing the L pole}

\item{R_vars}{a vector of variable names in x representing the R pole}

\item{N_vars}{a string containing a variable name in x representing N (total number of coded sentences)}

\item{method}{position scaling method: 'lowe' (default), 'budge', or 'kimfording'}
}
\description{
Calculate manifesto positions for each observation
}
