% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{manifesto_versions}
\alias{manifesto_versions}
\title{List available Manifesto Project Dataset versions}
\usage{
manifesto_versions()
}
\value{
A data.frame with columns 'id' and 'name'
}
\description{
List available Manifesto Project Dataset versions
}
