% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_data.R
\name{parlgov_linktable}
\alias{parlgov_linktable}
\title{Create a lookup table from parlgov data}
\usage{
parlgov_linktable(
  data_url = "https://partyfacts.herokuapp.com/download/external-parties-csv/"
)
}
\description{
Create a lookup table from parlgov data
}
