% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_data.R
\name{partyfacts_linktable}
\alias{partyfacts_linktable}
\title{Create a partyfacts lookup table}
\usage{
partyfacts_linktable(
  data_url = "https://partyfacts.herokuapp.com/download/external-parties-csv/",
  ignore_years = FALSE
)
}
\description{
Create a partyfacts lookup table
}
