% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview.R
\name{plotStructuralOverview}
\alias{plotStructuralOverview}
\title{plotStructuralOverview}
\usage{
plotStructuralOverview(structural_classifications)
}
\arguments{
\item{structural_classifications}{should be a tibble containning structural classifications. See details below}
}
\description{
plot a matrix structural overview
}
\details{
The tibble columns should firstly have MF, Adduct then followed by the structural classification taxonomic levels.
}
