% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eu_composition.R
\name{subset_to_eu}
\alias{subset_to_eu}
\title{Return all EU observations for a particular date}
\usage{
subset_to_eu(
  x,
  date = NA,
  country_id_var = attr(x, "country_id_var"),
  country_id_linktable = NA
)
}
\arguments{
\item{x}{the dataset to subset}

\item{date}{specified date for EU composition}

\item{country_id_var}{name of the country id variable in x}
}
\description{
Return all EU observations for a particular date
}
\section{TODO}{

- implement linktable country id handling (incl. varying id_vars for x and linktable?)
- add option to subset to eu when only supplying party_id (link and then use parlgov to obtain country_id and call recursive?)
}

