% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values.R
\name{values}
\alias{values}
\title{Retrieve positional values from a dataset}
\usage{
values(x, ...)
}
\arguments{
\item{x}{a dataset or composition object}

\item{...}{additional arguments passed to methods}
}
\description{
Retrieve positional values from a dataset
}
