% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eu_council_voting_weights.R
\name{council_voting_weights}
\alias{council_voting_weights}
\title{Calculate voting weight in the Council for a specific date}
\usage{
council_voting_weights(country_id, date, type = "bi_normal")
}
\arguments{
\item{country_id}{vector of ParlGov country ids}

\item{date}{a date in character format}
}
\description{
Calculate voting weight in the Council for a specific date
}
