% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{download_parlgov}
\alias{download_parlgov}
\title{Download a ParlGov CSV file}
\usage{
download_parlgov(
  base_url = "https://parlgov.org/data/parlgov-development_csv-utf-8/",
  type
)
}
\arguments{
\item{base_url}{base URL for ParlGov CSV files}

\item{type}{ParlGov view type: 'election', 'cabinet', or 'party'}
}
\value{
A data.frame
}
\description{
Download a ParlGov CSV file
}
