% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{download_parlgov_commission}
\alias{download_parlgov_commission}
\title{Download the Doering commissioner dataset}
\usage{
download_parlgov_commission(
  url =
    "https://www.parlgov.org/data/parlgov-development_csv-utf-8/external_commissioner_doering.csv"
)
}
\arguments{
\item{url}{URL or file path to the commissioner CSV}
}
\value{
A data.frame
}
\description{
Download the Doering commissioner dataset
}
