% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{parlgov_commission_dataset}
\alias{parlgov_commission_dataset}
\title{Create a commissioner dataset from the Doering/ParlGov data}
\usage{
parlgov_commission_dataset(
  url =
    "https://www.parlgov.org/data/parlgov-development_csv-utf-8/external_commissioner_doering.csv",
  update = TRUE
)
}
\arguments{
\item{url}{URL or file path to the commissioner CSV}

\item{update}{If TRUE (default), applies known corrections and additions
to the Doering dataset (e.g. VDL I commission personnel changes).}
}
\value{
A dataset S3 object of type 'parlgov_commission'
}
\description{
Create a commissioner dataset from the Doering/ParlGov data
}
