% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{parlgov_dataset}
\alias{parlgov_dataset}
\title{Create a ParlGov dataset object}
\usage{
parlgov_dataset(
  base_url = "https://parlgov.org/data/parlgov-development_csv-utf-8/",
  type
)
}
\arguments{
\item{base_url}{base URL for ParlGov CSV files}

\item{type}{ParlGov view type: 'election', 'cabinet', or 'party'}
}
\value{
A dataset S3 object
}
\description{
Create a ParlGov dataset object
}
