% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_statistics.R
\name{position_statistic}
\alias{position_statistic}
\title{Calculate (weighted) positions statistics}
\usage{
position_statistic(x, w, measure, na.rm = TRUE)
}
\arguments{
\item{x}{a vector of positions}

\item{w}{a vector of weights the same length as x}

\item{measure}{the statistics to compute, 'mean', 'median', 'sd'}

\item{na.rm}{whether to exclude `NA` values from the calculation}
}
\description{
Calculate (weighted) positions statistics
}
