% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values.R
\name{values.default}
\alias{values.default}
\title{Retrieve raw positional values in a dataset for given ID(s)}
\usage{
\method{values}{default}(
  input,
  date = NULL,
  data,
  value_vars,
  tolerance_lower = 365 * 6,
  tolerance_upper = 365 * 2,
  id_var = attr(data, "party_id_var"),
  date_var = attr(data, "date_var"),
  return_id_var = FALSE,
  additional_vars = c()
)
}
\arguments{
\item{input}{vector of IDs matching the ids in the target data}

\item{date}{date for the values to retrieve when there are repeated measurements}

\item{data}{target dataset dataframe}

\item{value_vars}{names of the value variables to retrieve form the target data}

\item{tolerance_lower}{Lower tolerance for dates in the target data (see \code{\link{determine_match_date}})}

\item{tolerance_upper}{Upper tolerance for dates in the target data (see \code{\link{determine_match_date}})}

\item{id_var}{name of the ID variable in the target data}

\item{date_var}{name of the date variable in the target data}

\item{return_id_var}{whether to include the id variable in the return}

\item{additional_vars}{additional variables from the values dataset to include in return}
}
\value{
a dataframe of the values for the values_vars from the target data for the given input
}
\description{
Retrieve raw positional values in a dataset for given ID(s)
}
